move utf parts into src dir
