+++ lib/src/utf_stream.dart
