+++ lib/utf.dart
@@ -19,7 +19,7 @@ import 'src/util.dart';
 export 'src/constants.dart';
 export 'src/utf_16_code_unit_decoder.dart';
 
-part "src/utf/utf_stream.dart";
-part "src/utf/utf8.dart";
-part "src/utf/utf16.dart";
-part "src/utf/utf32.dart";
+part "src/utf_stream.dart";
+part "src/utf8.dart";
+part "src/utf16.dart";
+part "src/utf32.dart";
